Add cmake files and first source file.
The cmake file consists of main file and file for listing the sources.
It is set to create a shared library.
Version handling of the source file does not work so far.

The source file is for io.
So far it contains a function to return an unused file unit number and
an efit class.
The code compiles but has not been tested so far.