Remove dependence on gbpi_mod.
Was used for handling error information?
Instead the derivs routine is now expected to take an additional
argument, which returns the error-status.

Note: in odeint was an error check after a call to rkqs, while the
routine (or the one that is called therein) does the error handling
(i.e. reporting) themselfes. As this is redundant, and the error status
is not passed back, the check has been removed.