Remove mpi related code.
Touches issue #1

Remove the mpi related code from arnoldi. This mean the direct use of
mpi (which was hidden behind the preprocessor) as well as the indirect
use via MyMPILib (her represented by mpro object).
The caller is now responsible for parallelization.
Not sure the code would work in parallel as it is now, espacially the
initialization.