Add test to execute test_arnoldi.x.
For this purpose a wrapper shell script has been writen for two reasons.
For one, fortran programs do not seem to return a value different from
zero when encountering a 'stop', while a nonzero return value on error
is expected by ctest. Second, the python script test_arnoldi.py needs to
be executed, before test_arnoldi.x can be run, as it will create the
required references.
The first can and is circumvented with setting the property
FAIL_REGULAR_EXPRESSION (looking for 'STOP');

Note: the wrapper might be unnecessary, as test_arnoldi.py &&
test_arnoldi.x might do the job.