Add module system_utility.
Main part is function for getting resource usage, based on the
c-function getrusage.
Unfortunately it does not work yet. The code compiles but the call of
getrusage returns non-zero, and all the fields of the structure are
zero.
Guess is that the structure defined in fortran has not the required
type(s).