Make getrusage work.
First version did not work, probably because the structure did not
match. If this is correct, then it should not be possible to call the
c-function directly from fortran as the c-structure uses unsigned
numbers which are not compatible with any fortran type.
To circumvent this, a wrapper for the function and the structure have
been written (in c).
The test is also added to the list of tests.