Introduce sorting in stretch_coords.
Make some changes in subroutine stretch_coords to move closer to the
neo-2 version of the subroutine.
Thus code for sorting according to theta is introduced.
This should only result in changes at most in case the points in the
input file used by stretch_coords are not in increasing order with
respect to theta (and even then it might make no difference at all).