Add 'accuracy' as attribute for float fields.
Add attribute 'accuracy' as optional input to subroutines that deal with
floating point fields (i.e. double and complex).

Setting is done by specialized attribute setting subroutine, which calls
the default one for comment and unit.

Parameter had to be one element array, because the corresponding hdf5
function expects an array, and not only a single value.