Add tool h5merge.
Thought to fit within the framework of this library, and so far not
under version control.

Note that the cmake file is supposed to be a first version. It should
also be usable with the main make file. At the moment this is not the
case due to dependency issues. The h5tools library needs to be present
to be found when configuration is done, which is not the case for first
configuration.
Also the search paths for the libraries contain hard-coded Build-folder
names.