Python script, convert neo.in to fortran namelist
Add a python script which can convert a neo.in file to a fortran
namelist and append it to another file.

Note: is already available in a neo-2 branch? (not sure if the commit is
already pushed), but is thought to fit better here, as this might also
be relevant for other codes (e.g. neo-rt).