First version of efit_to_boozer subroutine.
Add a first version of the efit_to_boozer subroutine. It compiles, but
is not working due to missing subroutines and possibly also a module for
data exchange.
The missing subroutines are (no guarantee that the list is complete):
- flint_for_Boozer
- spline_magdata_in_symfluxcoord
- magdata_in_symfluxcoord_ext
- field_eq
Of course, these are expected to have in turn also dependencies on other
modules and subroutines.