Bugfix: two variables had the same name.
Two variables in conversion from vmec to booozer routines had the same
name.
This was not a problem in the function version as the two variables
where used one after the other and each new flux surface would be
another function call wich resets the variables.
In the method version this is no longer the case, and thus needs to be
fixed.