Bugfix: value of nsurf was one to large.
Now actually refers to number of flux surfaces stored within the class.
This is one less as the number of surfaces in the netcdf file, as the
first flux surface from the netcdf file contains only zeros.
Also the value is now read only once, and the upper loop bound has been
adapted.