Add method for extracting modes.
Add a method for extracting modes to BoozerFile.
This is done to be able to do the complete conversion process in python.
So far for this an octave function had to be used.

The method was tested with a boozer file and gave the same results as
the octave function, minus some formatting (number of whitespaces in
some cases, exponent or not in some cases).

It is not yet clear that the python version has the same capabilites as
the octave version. This espacialy refers to the 'cos_harm' parameter
and the internal switch of the octave version.