Add extraction to main function.
Add extraction of modes to main function of boozer.py.
Saving of file with all modes has so far not been removed. This is a
safety measure, in case something goes wrong. This might be because not
all features of the octave extraction function have been implemented in
the python version.