Reduce order of Lagrange interpolation.
High order seems to make problems. The recent increase in slice size
seems to have made this even worse (by also affecting the order of the
Lagrange polynomial?). Thus now size of slice is reduced to three points
at most.
Reduction to linear order should be sufficient according to Sergei, as
vmec only uses linear interpolation in radius. But as derivatives are
required, this might lead to problems due to non-continous derivatives.