Bugfixes: import and format of members.
Import for getHeaderDataVMEC did not work anymore when
convert_vmec_to_boozer was used on its own, thus changed to 'from libneo
import ...' variant.
From main function the import statement was removed as it is no longer
needed.
2D Members of BoozerFile class (like m, n, or bmnc) are now lists of
lists after convert_vmec_to_boozer, as is expected in other parts of the
code.